+++ src/test/metrics.test.ts
@@ -15,6 +15,7 @@ const test = suite<{rig: WireitTestRig}>();
 test.before.each(async (ctx) => {
   try {
     ctx.rig = new WireitTestRig();
+    ctx.rig.env['WIREIT_LOGGER'] = 'metrics';
     await ctx.rig.setup();
   } catch (error) {
     // Uvu has a bug where it silently ignores failures in before and after,